requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
paho-mqtt>=1.6.1

# WordPress integration
requests>=2.26.0
python-jose>=3.3.0
python-multipart>=0.0.5

# Development dependencies
pytest>=6.2.5
black>=21.7b0
flake8>=3.9.2
mypy>=0.910
